src/functions/requirements.txt
azure-functions==1.21.3
azure-durable-functions==1.2.9
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
azure-search-documents==11.6.0
msgraph-beta-sdk>=1.5.0
pyodbc==5.1.0
requests==2.32.3
pandas==2.2.2
openpyxl==3.1.5
